MembersInvitationBackend/requirements.txt
# FastAPI and Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.17

# Database
sqlalchemy>=2.0.36
pg8000>=1.31.2
alembic>=1.14.0

# Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.2.1

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Testing
pytest>=8.3.0
